msdocs-python-flask-webapp-quickstart/requirements.txt
flask[async]
gunicorn
azure-identity
azure-storage-blob
aiohttp